% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{clean_labels}
\alias{clean_labels}
\title{Create a dataframe of labels for the end of the line graph}
\usage{
clean_labels(df)
}
\arguments{
\item{df}{A dataframe generated by download_game_data()}
}
\value{
A dataframe for passing to the labels argument of geom_text()
}
\description{
Create a dataframe of labels for the end of the line graph
}
