% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_game_data}
\alias{download_game_data}
\title{Download data from the Skanderbeg API}
\usage{
download_game_data(saves, api_key, values)
}
\arguments{
\item{saves}{A vector of Skanderbeg save strings (ID at the end of the URL)}

\item{api_key}{A string with your Skanderbeg API key}

\item{values}{A vector of values from the Skanderbeg getCountryData API}
}
\value{
A dataframe
}
\description{
Download data from the Skanderbeg API
}
