% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{extract_colours}
\alias{extract_colours}
\title{Extract tag colours from a dataframe to create a colour scale}
\usage{
extract_colours(df)
}
\arguments{
\item{df}{A dataframe generated by download_game_data()}
}
\value{
A character vector for passing to scale_colour_manual()
}
\description{
Extract tag colours from a dataframe to create a colour scale
}
