% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_session_history}
\alias{plot_session_history}
\title{Plot the session history for a metric}
\usage{
plot_session_history(df, metric, metric_string, file_path)
}
\arguments{
\item{df}{}

\item{metric_string}{}

\item{file_path}{}
}
\value{
A saved plot of the per-tag session history for that metric
}
\description{
Plot the session history for a metric
}
