% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{remove_tags}
\alias{remove_tags}
\title{Remove unwanted tags from the data}
\usage{
remove_tags(df, tag_removal_df)
}
\arguments{
\item{df}{A dataframe generated by download_game_data()}

\item{tag_removal_df}{A dataframe with two columns: tag and remove_until}
}
\value{
A dataframe
}
\description{
Remove unwanted tags from the data
}
