% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{save_plot}
\alias{save_plot}
\title{Save the session plot to disk in phone format}
\usage{
save_plot(plot, file_path, file_name)
}
\arguments{
\item{plot}{A plot object created by plot_session_history()}

\item{file_path}{A folder to save the plot to}

\item{file_name}{The name of the file to save as .png}
}
\description{
Save the session plot to disk in phone format
}
