% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{switch_tags}
\alias{switch_tags}
\title{Update data to reflect tag switches}
\usage{
switch_tags(df, tag_switch_df)
}
\arguments{
\item{df}{A dataframe generated by download_game_data()}

\item{tag_switch_df}{A dataframe with two columns: tag and new_tag}
}
\value{
A dataframe
}
\description{
Update data to reflect tag switches
}
